Added client-side data fetch

+++ pages/api/supplier.js
@@ -1,23 +1,25 @@
 import { pool } from '../../helpers/constants';
+import auth0 from '../../lib/auth0';
 import { v4 as uuid} from 'uuid';
 
 export default (req, res) => {
-  pool.getConnection((err, connection) => {
+  pool.getConnection(async (err, connection) => {
     if (err) {
       console.error("Connection error: " + err);
       res.status(500).send(err);
     }
+    const session = await auth0.getSession(req);
 
     switch (req.method) {
       case "GET": {
-        let { user, cols, lowerCase } = req.query;
+        let { cols, lowerCase } = req.query;
         if (!cols) cols = '*';
-        const sql = `SELECT ${cols} FROM suppliers WHERE user='${user}'`;
+        const sql = `SELECT ${cols} FROM suppliers WHERE user='${session ? session.user.name : "guest"}'`;
+
         connection.query(sql, (err, rows) => {
           if (err) {
             console.error("Get results error: " + err);
             res.status(500).send(err);
-            return;
           }
 
           const suppliers = rows.map(supplier => 
@@ -25,18 +27,16 @@ export default (req, res) => {
               { ...supplier, name: supplier.name.toLowerCase() }:
               supplier)
           res.status(200).send({suppliers: suppliers.sort((a, b) => a.name.localeCompare(b.name))});
-          
         });
 
         return
       }
 
       case "POST": {
-        const { user } = req.query;
         const { name, companyId, address, phone, email, comments } = req.body.data;
         const sql = 
           `INSERT INTO suppliers (_id, name, company_id, address, phone, email, comments, user)
-          VALUES ('${uuid()}', '${name}', '${companyId}', '${address}', '${phone}', '${email}', '${comments}', '${user}')`;
+          VALUES ('${uuid()}', '${name}', '${companyId}', '${address}', '${phone}', '${email}', '${comments}', '${session ? session.user.name : "guest"}')`;
 
         connection.query(sql, err => {
             if (err) {
@@ -50,39 +50,61 @@ export default (req, res) => {
       }
 
       case "PUT": {
-        const { user, _id } = req.query;
+        const { _id } = req.query;
         const { name, companyId, address, phone, email, comments } = req.body.data;
-        const sql = 
-          `UPDATE suppliers
-          SET name='${name}', company_id='${companyId}', address='${address}',
-            phone='${phone}', email='${email}', comments='${comments}'
-          WHERE _id='${_id}'`;
 
-        connection.query(sql, err => {
-            if (err) {
-              console.error("Insert to db error: " + err);
-              res.status(500).send(err);
-            }
+        let sql = `SELECT user FROM suppliers WHERE _id='${_id}'`;
+        connection.query(sql, (err, rows) => {
+          if (err) {
+            console.error("db error: " + err);
+            res.status(500).send(err);
+          }
+          if (rows[0].user == (session ? session.user.name : "guest")) {
+            sql = 
+              `UPDATE suppliers
+              SET name='${name}', company_id='${companyId}', address='${address}',
+                phone='${phone}', email='${email}', comments='${comments}'
+              WHERE _id='${_id}'`;
+    
+            connection.query(sql, err => {
+                if (err) {
+                  console.error("Insert to db error: " + err);
+                  res.status(500).send(err);
+                }
+            });
+            res.status(200).send('Success');
+
+          } else res.status(401).send("Unauthorized");
         });
 
-        res.status(200).send('Success');
-        return;
+        return
       }
 
       case "DELETE": {
-        const { user, _id } = req.query;
-        const sql = 
-          `DELETE FROM suppliers
-          WHERE _id='${_id}'`;
+        const { _id } = req.query;
+        let sql = `SELECT user FROM suppliers WHERE _id='${_id}'`;
 
-        connection.query(sql, err => {
-            if (err) {
-              console.error("Insert to db error: " + err);
-              res.status(500).send(err);
-            }
+        connection.query(sql, (err, rows) => {
+          if (err) {
+            console.error("Insert to db error: " + err);
+            res.status(500).send(err);
+          }
+          if (rows[0].user == (session ? session.user.name : "guest")) {
+            sql = 
+              `DELETE FROM suppliers
+              WHERE _id='${_id}'`;
+
+            connection.query(sql, err => {
+              if (err) {
+                console.error("Delete from db error: " + err);
+                res.status(500).send(err);
+              }
+            });
+            res.status(200).send('Success');
+
+          } else res.status(401).send("Unauthorized");
         });
-
-        res.status(200).send('Success');
+        
         return;
       }
